improved styling

+++ lib/widgets/chart.dart
@@ -33,7 +33,7 @@ class Chart extends StatelessWidget {
 
   double get totalSpending {
     return groupedTransactionValues.fold(0.0, (sum, item) {
-      return sum + (item['amount'] as int);
+      return sum + (item['amount'] as double);
     });
   }
